+++ CHANGELOG.md
@@ -14,8 +14,8 @@ and this project adheres to
 - Constructors of `Ability`, `Race` and `Region`.
 - Parameters of `validate()`.
 - Speed up `Game` construction.
-- `Game` constructor now validates the number of `races` in assets
-    and may raise `InvalidAssets` exception.
+- `Game` constructor now validates the number of `races` and `abilities`
+    in assets, and may raise `InvalidAssets` exception.
 - `assets/tiny.json` now contains more `races` to pass the checks.
 
 ### Removed